GO-2318 Use space.Do to update type SB
